backend/ms_index/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
pydantic==2.5.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
gitpython==3.1.40
javalang==0.15.1
numpy==1.24.3
python-multipart==0.0.6